+++ code/header/model/energy_storage_model/estr0.h
@@ -107,16 +107,16 @@ class ESTR0 : public ENERGY_STORAGE_MODEL
 
         LEAD_LAG_BLOCK active_lead_lag_1, active_lead_lag_2;
         PID_BLOCK active_pid_block;
-        float Pmax;
+        double Pmax;
         FIRST_ORDER_BLOCK active_power_filter;
         INTEGRAL_BLOCK energy_state_block;
         double E0;
-        float Enominal;
+        double Enominal;
 
-        float Effin, Effout;
+        double Effin, Effout;
         LEAD_LAG_BLOCK reactive_lead_lag_1, reactive_lead_lag_2;
         INTEGRAL_BLOCK reactive_integral_block;
-        float Dq;
+        double Dq;
 
 
 };
