+++ code/header/model/sg_models/turbine_governor_model/IEEEG1SB.h
@@ -106,23 +106,23 @@ class IEEEG1SB : public TURBINE_GOVERNOR_MODEL
         void copy_from_const_model(const IEEEG1SB& model);
 
         LEAD_LAG_BLOCK droop;
-        float Tservo;
-        float Uo, Uc;
+        double Tservo;
+        double Uo, Uc;
         INTEGRAL_BLOCK servo_motor;
         FIRST_ORDER_BLOCK delayer1;
-        float K1, K2;
+        double K1, K2;
         FIRST_ORDER_BLOCK delayer2;
-        float K3, K4;
+        double K3, K4;
         FIRST_ORDER_BLOCK delayer3;
-        float K5, K6;
+        double K5, K6;
         FIRST_ORDER_BLOCK delayer4;
-        float K7, K8;
+        double K7, K8;
         PI_BLOCK boiler_PI;
         FIRST_ORDER_BLOCK fuel_producer;
         FIRST_ORDER_BLOCK fuel_delayer1;
         FIRST_ORDER_BLOCK fuel_delayer2;
         INTEGRAL_BLOCK drum;
-        float Kb;
+        double Kb;
 
 };
