+++ code/header/model/sg_models/turbine_governor_model/IEEEG2.h
@@ -66,8 +66,8 @@ class IEEEG2 : public TURBINE_GOVERNOR_MODEL
 
         FIRST_ORDER_BLOCK droop;
         LEAD_LAG_BLOCK tuner;
-        float Pmax, Pmin;
-        float K3, K4;
+        double Pmax, Pmin;
+        double K3, K4;
         LEAD_LAG_BLOCK water_hammer;
 };
